Codechange: Move AI settings back to ai_gui
This also opens GS's own config window from Debug window, and reverts caption fixes.
