+++ src/company_cmd.cpp
@@ -922,7 +922,7 @@ CommandCost CmdCompanyCtrl(DoCommandFlag flags, CompanyCtrlAction cca, CompanyID
 	}
 
 	InvalidateWindowClassesData(WC_GAME_OPTIONS);
-	InvalidateWindowClassesData(WC_SCRIPT_SETTINGS);
+	InvalidateWindowClassesData(WC_AI_SETTINGS);
 	InvalidateWindowClassesData(WC_SCRIPT_LIST);
 
 	return CommandCost();
